requirements.txt
requests
asyncio
kafka-python
numpy
pandas
datetime